requirements.txt
docxtpl==0.19.1
pandas==1.3.3
python_dateutil==2.8.2
python_docx==1.1.2
python-docx
lxml
python-dateutil